fixed remaining bugs. Implementing 'hover over order number to view order details' feature

+++ update_order.php
@@ -9,6 +9,11 @@ if(!isset($_SESSION["user"]))
 require("settings.php");
 $product_quantities = $_POST["products"];
 $orderNum = $_POST['order_number'];
+$user = $_SESSION['user'];
+
+$query = "SELECT money FROM user WHERE id = '{$user->id}'";
+$result = $db_con->query($query);
+$user_money = mysqli_fetch_object($result);
 
 $query = "SELECT SUM(p.cost * op.quantity) AS cost
 			FROM order_product op
@@ -21,20 +26,35 @@ $new_price = 0;
 for ($i=0; $i < count($product_quantities); $i++) { 
 	$new_price += $product_quantities[$i]['product_cost'] * $product_quantities[$i]['product_quantity'];
 }
-$new = ($new_price - $current_amount->cost);
+$new = ($current_amount->cost - $new_price);
 
-var_dump("new price = " . $new_price);
-var_dump("current amount = " . $current_amount->cost);
-var_dump("new = " . $new);
-die();
+$positive = 0;
+if ($new < 0) // increasing quantities = deduct money
+{
+	$positive = $new / -1;
+	$new_money = $user_money->money - $positive;
+	
+}
+else // decreasing quantities = refund money
+{
+	$new_money = $user_money->money + $new;
+}
 
+if ($new_money < 0)
+{
+	die("You do not have enough money for the update.");
+}
+else
+{
+	$statement = $db_con->prepare("UPDATE user SET money = ? WHERE id = ?");
+	$statement->bind_param("ii", $new_money , $user->id);
+	$statement->execute();
 
+	$query = "DELETE FROM order_product WHERE order_ref = '{$orderNum}'";
+	$result = $db_con->query($query);
 
-$query = "DELETE FROM order_product WHERE order_ref = '{$orderNum}'";
-$result = $db_con->query($query);
-
-$query = "UPDATE user_order SET order_update = NOW() WHERE id = '{$orderNum}'";
-$result = $db_con->query($query);
+	$query = "UPDATE user_order SET order_update = NOW() WHERE id = '{$orderNum}'";
+	$result = $db_con->query($query);
 
 	if($statement = $db_con->prepare("INSERT INTO order_product VALUES(?, ?, ?)"))
 	{
@@ -47,6 +67,7 @@ $result = $db_con->query($query);
 			$statement->execute();
 		}
 	}
+}
 
 	// $query = "SELECT u.firstname, uo.id, SUM(p.cost * op.quantity) AS value, uo.order_date, uo.order_update, s.label AS status
 	// 	FROM user_order uo
@@ -67,4 +88,4 @@ $result = $db_con->query($query);
 	// 	}
 		
 	// }
-die(json_encode($order));
+die($new_money);
